+++ sdk/metric/controller/basic/controller_test.go
@@ -44,7 +44,7 @@ func getMap(t *testing.T, cont *controller.Controller) map[string]float64 {
 	require.NoError(t, cont.ForEach(
 		func(_ instrumentation.Library, reader export.Reader) error {
 			return reader.ForEach(
-				export.CumulativeExportKindSelector(),
+				aggregation.CumulativeTemporalitySelector(),
 				func(record export.Record) error {
 					return out.AddRecord(record)
 				},
@@ -114,7 +114,7 @@ func TestControllerUsesResource(t *testing.T) {
 	}
 	for _, c := range cases {
 		t.Run(fmt.Sprintf("case-%s", c.name), func(t *testing.T) {
-			sel := export.CumulativeExportKindSelector()
+			sel := aggregation.CumulativeTemporalitySelector()
 			exp := processortest.New(sel, attribute.DefaultEncoder())
 			cont := controller.New(
 				processor.NewFactory(
@@ -144,7 +144,7 @@ func TestStartNoExporter(t *testing.T) {
 	cont := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			aggregation.CumulativeTemporalitySelector(),
 		),
 		controller.WithCollectPeriod(time.Second),
 		controller.WithResource(resource.Empty()),
@@ -213,7 +213,7 @@ func TestObserverCanceled(t *testing.T) {
 	cont := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			aggregation.CumulativeTemporalitySelector(),
 		),
 		controller.WithCollectPeriod(0),
 		controller.WithCollectTimeout(time.Millisecond),
@@ -245,7 +245,7 @@ func TestObserverContext(t *testing.T) {
 	cont := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			aggregation.CumulativeTemporalitySelector(),
 		),
 		controller.WithCollectTimeout(0),
 		controller.WithResource(resource.Empty()),
@@ -277,7 +277,7 @@ type blockingExporter struct {
 func newBlockingExporter() *blockingExporter {
 	return &blockingExporter{
 		exporter: processortest.New(
-			export.CumulativeExportKindSelector(),
+			aggregation.CumulativeTemporalitySelector(),
 			attribute.DefaultEncoder(),
 		),
 	}
@@ -295,11 +295,11 @@ func (b *blockingExporter) Export(ctx context.Context, res *resource.Resource, o
 	return err
 }
 
-func (*blockingExporter) ExportKindFor(
+func (*blockingExporter) TemporalityFor(
 	*metric.Descriptor,
 	aggregation.Kind,
-) export.ExportKind {
-	return export.CumulativeExportKind
+) aggregation.Temporality {
+	return aggregation.CumulativeTemporality
 }
 
 func TestExportTimeout(t *testing.T) {
@@ -307,7 +307,7 @@ func TestExportTimeout(t *testing.T) {
 	cont := controller.New(
 		processor.NewFactory(
 			processortest.AggregatorSelector(),
-			export.CumulativeExportKindSelector(),
+			aggregation.CumulativeTemporalitySelector(),
 		),
 		controller.WithCollectPeriod(time.Second),
 		controller.WithPushTimeout(time.Millisecond),
@@ -356,7 +356,7 @@ func TestExportTimeout(t *testing.T) {
 
 func TestCollectAfterStopThenStartAgain(t *testing.T) {
 	exp := processortest.New(
-		export.CumulativeExportKindSelector(),
+		aggregation.CumulativeTemporalitySelector(),
 		attribute.DefaultEncoder(),
 	)
 	cont := controller.New(
@@ -435,7 +435,7 @@ func TestCollectAfterStopThenStartAgain(t *testing.T) {
 
 func TestRegistryFunction(t *testing.T) {
 	exp := processortest.New(
-		export.CumulativeExportKindSelector(),
+		aggregation.CumulativeTemporalitySelector(),
 		attribute.DefaultEncoder(),
 	)
 	cont := controller.New(